added /modify/ policy to authorize only blog owner to taking action + implements HasMiddleware controller for authenticated users only

+++ app/Http/Controllers/PostController.php
@@ -4,10 +4,20 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth', except: ['show', 'index']),
+        ];
+    }
+
     /**
      * Display posts.
      */
@@ -21,11 +31,11 @@ class PostController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
+     * Display the blog
      */
-    public function create()
+    public function show(Post $post)
     {
-        //
+        return view('posts.show', ["post" => $post]);
     }
 
     /**
@@ -33,6 +43,8 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        // we won't need Gate to check ownernship because auth middleware is present
+
         // validate
         $fields = $request->validate(
             [
@@ -49,18 +61,13 @@ class PostController extends Controller
     }
 
     /**
-     * Display the blog
-     */
-    public function show(Post $post)
-    {
-        return view('posts.show', ["post" => $post]);
-    }
-
-    /**
      * Show the page to edit the blog.
      */
     public function edit(Post $post)
     {
+        // use modify policy to check owner
+        Gate::authorize('modify', $post);
+
         return view('posts.edit', ["post" => $post]);
     }
 
@@ -69,12 +76,19 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        // use modify policy to check owner
+        Gate::authorize('modify', $post);
+
+        // Validate
         $fields = $request->validate([
             "title" => ["required", "max:255"],
             "body" => ["required"]
         ]);
+
+        // Update
         $post->update($fields);
 
+        // Redirect to dashboard
         return redirect()->route('dashboard')->with('update', 'Your post has been updated');
     }
 
@@ -83,8 +97,13 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        // use modify policy to check owner
+        Gate::authorize('modify', $post);
+
+        // Delete
         $post->delete();
 
+        // Redirect back
         return back()->with('delete', 'your post was deleted!');
     }
 }
